+++ HISTORY.rst
@@ -3,6 +3,11 @@
 History
 -------
 
+1.0.1 (2013-03-20)
+++++++++++++++++++
+- Fixed a bug where classes not extending from the Python exception hierarchy could slip through
+- Update test suite for custom Python exceptions
+
 1.0.0 (2013-01-21)
 ++++++++++++++++++
 - First stable, tested version now exists